python/requirements.txt
notebooklm-mcp-cli>=0.5.0
